Changed Game Class to Follow State Design Pattern
I took out the time delay in character (which prevented constant
movement) and replaced it with an overall system key_repeat setting (now
events are updated every 100 ms when a key is held down with the held
down key). This just made it easier to write the Game class code.